Preventing double space when accepting intellisense suggestion.
This happens when a register from intellisense is selected as second argument for example. First `Z80TypingFormatter` inserts a space after first argument (if this is configured in options). Then `Z80CompletionProposer` inserts another one.

Interestingly: if a label is selected, then double space is not inserted. But behavior is consistent for both with these changes - formatter takes care of establishing the spacing before completion proposer is executed.