Improving experience when accepting intellisense items.
First of all, instead of just enter, tab is also accepted as a valid commit character. Then, for first argument, comma is accepted on top, and for second argument a space. This allows more natural and fluid autocomplete experience.

Secondly, newline is only added to the end of autocompletion if `formatOnType` is **not** used. This prevents a newline being injected into the text after accepting autocompletion. Again, this feels more natural since frequently user wants to append a comment after a command for example.